make datatypes smaller

+++ libraries/Encoder.sol
@@ -2,7 +2,7 @@
 pragma solidity =0.8.26;
 
 library DepositEncoder {
-    uint constant ENCODING_VERSION = 1;
+    uint8 constant ENCODING_VERSION = 1;
 
     enum DEPOSIT_TYPE {
         REPO,
@@ -11,8 +11,8 @@ library DepositEncoder {
 
     function encode(
         DEPOSIT_TYPE depositType,
-        uint         repoId,
-        uint         timestamp
+        uint64       repoId,
+        uint64       timestamp
     ) internal pure returns (bytes memory) {
         return abi.encode(ENCODING_VERSION, depositType, repoId, timestamp);
     }
